Tests/LibWeb/Text/input/css/serialize-at-font-face-missing-font-family.html

<!DOCTYPE html>
<html>
    <style>
        @font-face {
            src: url("./Foo.woff");
        }
    </style>
    <script src="../include.js"></script>
    <script>
        test(() => {
            println(document.styleSheets[0].cssRules[0].cssText);
        });
    </script>
</html>
