Tests/LibWeb/Text/expected/css/serialize-at-font-face-missing-font-family.txt

@font-face { src: url("./Foo.woff"); }
